Modularized everything! Tests are moved into test folder, linked list, node,functions, stack, and queue are in separate folders and imported.
