+++ code-challenges/linked-list/node.js
@@ -1,10 +1,11 @@
 'use strict';
 
-module.exports = Node;
 
 class Node {
   constructor(){
-    this.head = null;
-    this.tail = null;
+    this.value = null;
+    this.next = null;
   }
 }
+
+module.exports = Node;
